🐛 Handle start loading animation's completion properly
When an animation is stopped by the developer before the start loading
animation has finished, we reverse the start loading animation in
`stopLoading` method, and its completion for .start position will be called as
expected. In that completion method, we should not only deactivate
the constraints of the loading, we should remove the associated layout.